added map

+++ NerdDinner/Controllers/DinnersController.cs
@@ -44,13 +44,18 @@ namespace NerdDinner.Controllers
         public ActionResult Details(int id)
         {
             Dinner dinner = db.Dinners.Find(id);
-            //Dinner dinner = dinnersDb.Dinners.Find(id);
+            
             return dinner == null ? View("NotFound") : View("Details", dinner);
         }
 
+        [Authorize]
         public ActionResult Edit(int id)
         {
             Dinner dinner = db.Dinners.Find(id);
+
+
+            if (!dinner.IsHostedBy(User.Identity.Name))
+                return View("InvalidOwner");
 
             var countries = new List<string>();
             countries.Add("America");
@@ -62,12 +67,17 @@ namespace NerdDinner.Controllers
             return View(dinner);
         }
 
+        [Authorize]
         [AcceptVerbs(HttpVerbs.Post)]
         public ActionResult Edit(int id, FormCollection formValues)
         {
             // Retrieve existing dinner
             Dinner dinner = db.Dinners.Find(id);
 
+            if (!dinner.IsHostedBy(User.Identity.Name))
+            {
+                return View("InvalidOwner");
+            }
             UpdateModel(dinner);
 
             // Persist changes back to database
@@ -77,6 +87,7 @@ namespace NerdDinner.Controllers
             return RedirectToAction("Details", new { id = dinner.DinnerId });
         }
 
+        [Authorize]
         public ActionResult Create()
         {
 
@@ -96,25 +107,37 @@ namespace NerdDinner.Controllers
         }
 
         [AcceptVerbs(HttpVerbs.Post)]
+        [Authorize]
         public ActionResult Create(Dinner dinner)
         {
-            dinner.HostedBy = "some guy";
             if (ModelState.IsValid)
             {
+                
+                dinner.HostedBy = User.Identity.Name;
+
+                Rsvp rsvp = new Rsvp();
+                rsvp.AttendeeName = User.Identity.Name;
+                dinner.Rsvps.Add(rsvp);
+
                 db.Dinners.Add(dinner);
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
+
             return View(dinner);
         }
 
         public ActionResult Delete(int id)
         {
             Dinner dinner = db.Dinners.Find(id);
+
             if (dinner == null)
-            {
                 return View("NotFound");
-            }
+
+            if (!dinner.IsHostedBy(User.Identity.Name))
+                return View("InvalidOwner");
+            
+
             return View(dinner);
         }
 
@@ -122,10 +145,13 @@ namespace NerdDinner.Controllers
         public ActionResult Delete(int id, string confirmButton)
         {
             Dinner dinner = db.Dinners.Find(id);
-
+   
             if (dinner == null)
                 return View("NotFound");
 
+            if (!dinner.IsHostedBy(User.Identity.Name))
+                return View("InvalidOwner");
+            
             db.Dinners.Remove(dinner);
             db.SaveChanges();
